Move deepseek config to client.py and fix deepseek max_tokens upper limit
